If nodes are not registered at collaboration creation time, add option to register missing nodes later
