+++ vantage6-ui/src/app/pages/collaboration/read/collaboration-read.component.html
@@ -46,8 +46,19 @@
   <mat-card>
     <mat-card-header>
       <mat-card-title>{{ "collaboration-read.card-nodes.title" | translate }}</mat-card-title>
+      <button
+        actions
+        mat-flat-button
+        class="right-aligned-button"
+        color="primary"
+        *ngIf="canEdit"
+        (click)="onRegisterMissingNodes()"
+        >
+        <mat-icon>add</mat-icon>{{ "collaboration-read.card-nodes.add" | translate }}
+      </button>
     </mat-card-header>
     <mat-card-content>
+      <app-alert *ngIf="isMissingNodes()" class="node-alert" label="{{'collaboration-read.alert-missing-nodes' | translate}}"></app-alert>
       <div class="chip-container">
         <app-chip
           *ngFor="let node of collaboration.nodes"
