+++ vantage6-ui/src/app/services/node.service.ts
@@ -1,10 +1,15 @@
 import { Injectable } from '@angular/core';
 import { ApiService } from './api.service';
-import { APIKey, BaseNode, GetNodeParameters, Node, NodeCreate, NodeEdit, NodeLazyProperties } from '../models/api/node.model';
+import { APIKey, ApiKeyExport, BaseNode, GetNodeParameters, Node, NodeCreate, NodeEdit, NodeLazyProperties } from '../models/api/node.model';
 import { BaseCollaboration, Collaboration } from '../models/api/collaboration.model';
 import { OrganizationService } from './organization.service';
 import { Pagination } from '../models/api/pagination.model';
 import { getLazyProperties } from '../helpers/api.helper';
+import { BaseOrganization } from '../models/api/organization.model';
+import { FileService } from './file.service';
+import { TranslateService } from '@ngx-translate/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MessageDialogComponent } from '../components/dialogs/message-dialog/message-dialog.component';
 
 @Injectable({
   providedIn: 'root'
@@ -12,7 +17,10 @@ import { getLazyProperties } from '../helpers/api.helper';
 export class NodeService {
   constructor(
     private apiService: ApiService,
-    private organizationService: OrganizationService
+    private organizationService: OrganizationService,
+    private dialog: MatDialog,
+    private translateService: TranslateService,
+    private fileService: FileService
   ) {}
 
   async getNodes(parameters?: GetNodeParameters): Promise<BaseNode[]> {
@@ -60,4 +68,42 @@ export class NodeService {
   async resetApiKey(nodeID: string): Promise<APIKey> {
     return await this.apiService.postForApi<APIKey>(`/recover/node`, { id: nodeID });
   }
+
+  async registerNodes(collaboration: BaseCollaboration | Collaboration, organizations: BaseOrganization[]): Promise<void> {
+    const apiKeys: ApiKeyExport[] = [];
+    await Promise.all(
+      organizations.map(async (organization: BaseOrganization) => {
+        const node = await this.createNode(collaboration, organization.id);
+        if (node?.api_key) {
+          apiKeys.push({
+            organization: organization.name,
+            api_key: node.api_key
+          });
+        }
+      })
+    );
+    this.downloadApiKeys(apiKeys, collaboration.name);
+    this.alertApiKeyDownload();
+  }
+
+  private downloadApiKeys(api_keys: ApiKeyExport[], collaboration_name: string): void {
+    const filename = `API_keys_${collaboration_name}.txt`;
+    const text = api_keys.map((api_key) => `${api_key.organization}: ${api_key.api_key}`).join('\n');
+    this.fileService.downloadTxtFile(text, filename);
+  }
+
+  alertApiKeyDownload(): void {
+    this.dialog.open(MessageDialogComponent, {
+      data: {
+        title: this.translateService.instant('api-key-download-dialog.title'),
+        content: [
+          this.translateService.instant('api-key-download-dialog.create-message'),
+          this.translateService.instant('api-key-download-dialog.security-message')
+        ],
+        confirmButtonText: this.translateService.instant('general.close'),
+        confirmButtonType: 'default'
+      }
+    });
+  }
+
 }
